emotion-analysis/pyproject.toml
[tool.poetry]
name = "emotion_analysis"
version = "0.1.0"
description = ""
authors = ["fusedbloxxer <voidwarden@yahoo.com>"]
readme = "README.md"
packages = [{ include = "emotion_analysis" }]

[tool.poetry.dependencies]
python = "^3.11"
ipykernel = "^6.27.1"
seaborn = "^0.13.0"
flax = "^0.7.5"
matplotlib = "^3.8.2"
scipy = "^1.11.4"
numpy = "^1.26.2"
optax = "^0.1.7"
jax = "0.4.23"
jaxlib = {version = "0.4.23", source = "jax_cuda"}
python-dotenv = "^1.0.0"
transformers = "^4.36.2"
tqdm = "^4.66.1"
mlflow = "^2.9.2"
wordcloud = "^1.9.3"
scikit-learn = "^1.3.2"
ipywidgets = "^8.1.1"
av = "^11.0.0"
optuna = "^3.5.0"
datasets = "^2.16.1"
pillow = "^10.2.0"
pyyaml = "^6.0.1"
evaluate = "^0.4.1"
tensorflow-datasets = "^4.9.4"
torch = {version = "^2.1.2+cpu", source = "torch_cpu"}

[[tool.poetry.source]]
name = "PyPI"
priority = "default"

[[tool.poetry.source]]
name = "jax_cuda"
url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
priority = "explicit"

[[tool.poetry.source]]
name = "torch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
